requirements.txt
discord.py
mariadb
colorama
pyfiglet
